+++ packages/pigeon/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.27/Dart 3.6.
+
 ## 25.3.2
 
 * [dart] Fixes null pointer crashes/exceptions caused by premature finalization of Dart instances