+++ packages/vector_graphics/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.27/Dart 3.6.
+
 ## 1.1.18
 
 * Allow transition between placeholder and loaded image to have an animation.